+++ model/scratch/examples/scratchlgrsag/dataset.sh
@@ -1,0 +1,8 @@
+cat > dataset.csv << EOF
+f1,ans
+0.1,0
+0.7,1
+0.6,1
+0.2,0
+0.8,1
+EOF