+++ dffml/base.py
@@ -6,7 +6,6 @@ import abc
 import ast
 import copy
 import inspect
-import argparse
 import functools
 import contextlib
 import collections