+++ dffml/service/dev.py
@@ -5,7 +5,6 @@ import json
 import pydoc
 import shutil
 import asyncio
-import pathlib
 import getpass
 import tempfile
 import importlib