+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Python prompts `>>>` can now be enabled or disabled for easy copying of code into interactive sessions.
-- whitespace check now checks .rst and .md files too.
+- Whitespace check now checks .rst and .md files too.
 ### Changed
 - Renamed `"arg"` to `"plugin"`.