+++ CHANGELOG.md
@@ -7,7 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Python prompts `>>>` can now be enabled or disabled for easy copying of code into interactive sessions.
+<<<<<<< HEAD
 - Whitespace check now checks .rst and .md files too.
+=======
+- `GetMulti` operation which gets all Inputs of a given definition
+- Support for async generator operations
+>>>>>>> 422cdb3a3920005f5c244689bfce8294414da1ae
 ### Changed
 - Renamed `"arg"` to `"plugin"`.
 
@@ -21,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `validate` parameter in `Input` takes `Operation.instance_name`
 - New db source can utilize any database that inherits from `BaseDatabase`
 - Logistic Regression with SAG optimizer
+- Test tensorflow DNNEstimator documentation examples in CI
 - shouldi got an operation to run cargo-audit on rust code.
 - Moved all the downloads to tests/downloads to speed the CI test.
 - Test tensorflow DNNEstimator documentation exaples in CI