+++ CHANGELOG.md
@@ -8,8 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - IO operations demo
 - Python prompts `>>>` can now be enabled or disabled for easy copying of code into interactive sessions.
+- Whitespace check now checks .rst and .md files too.
+- `GetMulti` operation which gets all Inputs of a given definition
+- Python usage example for LogisticRegression and its related tests.
+- Support for async generator operations
 ### Changed
 - Renamed `"arg"` to `"plugin"`.
+### Fixed
+- Exporting plugins
 
 ## [0.3.6] - 2020-04-04
 ### Added
@@ -21,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `validate` parameter in `Input` takes `Operation.instance_name`
 - New db source can utilize any database that inherits from `BaseDatabase`
 - Logistic Regression with SAG optimizer
+- Test tensorflow DNNEstimator documentation examples in CI
 - shouldi got an operation to run cargo-audit on rust code.
 - Moved all the downloads to tests/downloads to speed the CI test.
 - Test tensorflow DNNEstimator documentation exaples in CI