+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Example CLI commands and Python code for `SLRModel`
 ### Changed
 - Renamed `"arg"` to `"plugin"`.
+- CSV source sorts feature names within headers when saving
 ### Fixed
 - Exporting plugins
 - Issue parsing string values when using the `dataflow run` command and