+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Python usage example for LogisticRegression and its related tests.
 - Support for async generator operations
 - Example CLI commands and Python code for `SLRModel`
+- `save` function in high level API to quickly save all given records to a
+  source
 ### Changed
 - Renamed `"arg"` to `"plugin"`.
 - CSV source sorts feature names within headers when saving