+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ability to configure sources and models for HTTP API from command line when
   starting server
 - Documentation page for command line usage of HTTP API
+- Source for parsing `.ini` file formats
 ### Changed
 - Renamed `"arg"` to `"plugin"`.
 - CSV source sorts feature names within headers when saving